requirements.txt
streamlit
gspread
pandas
google-auth
